breaking: v0.3.0 (#28)
## Breaking changes to keybinding logic and configuration

This release moves from partially unsupported keybindings for old terminals to a _modal-based_ action trigger (fixes #12) and uses more widely supported defaults. For the motivation, see #12 and the [Customization](https://isd-project.github.io/isd/customization/) page
for details. It is recommended to _delete_ the `~/.config/isd/config.yaml` file and regenerate it if required.

"Direct" systemctl actions need to be explicitly enabled (but are still supported!).

`isd` will also check for overlapping keybindings and raise an error to minimize unexpected behaviour of the tool.

## Updated command palette

The command palette includes more descriptive help texts and mentions the associated keyboard shortcut if available.
This should help with keybinding discovery.
The documentation now explicitly mentions how to quit the application (fixes #25 and solves #21). 

## New Logo

`isd` now comes with a new logo from a hobby artist who prefers to only be referenced as Denis. Thank you Denis!

## Pager fixes

`isd` uses a more widely used `less` argument configuration to also work with older versions of `less` (fixes #22)

## Templated services

`isd` now behaves more similarly to `systemd` when viewing templated service files. To better convey that they do not provide a `Preview` output (apart from `Cat`), these units are greyed out, as suggested in #16 (fixed #16).

## Basic error recovery

If the settings cannot be correctly loaded, the error message is shown and an error log is generated. This should help users with breaking changes in the future.