v0.4.0 (#38)
This release mainly focuses on improving the packaging side.
`isd` now provides a desktop file and uses a standard size for the icon (fixes #29).

This release will also be available on PyPI. The main reason for publishing
to PyPI is to better support packaging for distros as mentioned in #18.
Due to a naming clash on PyPI, the Python project has been renamed
to `isd-tui`. However, the project will continue to be named `isd`
and the executable will continue to be called `isd`!

The Python wheel only includes the necessary files, including
the `.desktop` file and icon.
The `nix` derivation propagates the `.desktop` file and the icon and makes them
available for Desktops of `nix` users.
Due to a bug in `isd-appimage`, identical `nixpkgs` pins cannot be used for `isd`
(#37 is blocked).

This release also introduces _EXPERIMENTAL_ arm support!
I do _not_ have access to an arm-based machine, and as such, I cannot actually
test the arm-based `AppImage`. The `AppImage` was created by cross-compiling. 
However, I cannot run a cross-platform `AppImage` (should fix #36).
Please let me know if it works for you or not 🙏

At least the arm-based `nix` derivation can be run with QEMU, and it seems to work.

PS: This release also includes a workaround for a `Zellij` bug (fixes #34).
Though upstream already has a bug-fix available until a new version has
been published, `isd` will work around the Zellij issue.