fix: improve service tab UX
Further rework of #48.
Now the tab headers are not focusable!
This means that switching to the tab headers themselves and moving them
left or right isn't possible anymore.
Only the "global" tab switching (and clicking) can change the preview
tab. This might cause a bit more friction for new users since they have
to discover this shortcut but that can be added later.
The new approach is a lot nicer once one is used to it, since one can
quickly switch between the tabs.