fix: clicking on unit now only highlights
Previously, clicking on a unit without a modifier selected the unit.
This was a surprising behavior for some users.
Now, clicking a unit only highlights the unit and is the same
as _moving_ to the it via keyboard navigation.
To select a unit the user has to press one of the modifier keys.