feat: Add support for `root` user to connect to `--user` services
Removed the dedicated screen from the previous commit and instead opted
for a quick local test before switching. This removes the need for the
user to configure anything and it should just work out-of-the-box and
still educate the user.
Now also allows the application to startup in the `user` mode when
called as the `root` user.

Fixes #30