fix: Authorization issues when interacting with `system` units
This should fix a major issue with newer (?) `systemd` versions. I am
not 100% sure if the issue came from an update or from a regression I
have introduced along the way. But now, `isd` only asks for a password
via a password agent if the `polkit` authorization mode is used.

As a result, executing `systemctl` actions on system units shouldn't
_freeze/crash_ the TUI anymore or pop up an _unnecessary_ password agent
UI.