deps: Update python dependencies and increase minimal textual version
`pytest` tests failed since a9df5ea73be6 when I updated the python
dependencies.
For one, the visual snapshots failed, as `textual` updated the coloring
for highlighted items.
-> Simply had to update the visual tests with `pytest --snapshot-update`
after ensuring that the `diff` still looked alright.

But an independent regression was that `query_exactly_one` failed to
match my nodes in the tests.
This was introduced in `textual` v3.0.0 where the `App.query` selector
matches the default/first screen
and not the active screen... I believe I only relied on this behavior in
my tests.
Still bumping the minimum textual dependency to be safe.