fix: Disable `cli` parsing
This will be necessary to create a "blessed" CLI interface and is a
first step towards resolving #64.
This features was more of a hack for internal use but it has
(rightfully) confused users in the past.
If anybody relied on this feature, they should set the environment
variables until the new CLI interface is implemented.