exp-feat: allow theme to be inherited from terminal
When using a modern terminal emulator such as kitty or ghostty, a new
theme called `terminal-derived-theme` can be selected. This theme will
infer the appropriate theme from the terminal palette colors. When an
unsupported terminal emulator is used, a notification will be raised and
the default selected.

This feature is EXPERIMENTAL. But it seems to be working nicely with
`ghostty` and `kitty`. Though in `kitty` there seems to be some
flickering.